:sparkles: Add openai help and commands
